Create README.md
Linked list is a data structure in which the data is not present in adjacent memory locations. The fundamental component of a linked list is a node which consists of data and address of next node. One node points to the next node and so on. The elements can be added and deleted from the front or the end of the linked list.